requirements.txt
joblib==1.5.2
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
uvicorn==0.38.0
fastapi==0.121.1
slowapi==0.1.9
PyJWT==2.10.1
python-dotenv==1.2.1
